feat(App_Folder): Migrate to App folder (Auth page and component page)
