+++ messages/vi.json
@@ -1,0 +1,50 @@
+{
+  "common": {
+    "Authentication failed Please try again": "Đăng nhập thất bại. Vui lòng thử  lại.",
+    "Create Account": "Tạo Tài Khoản",
+    "Email Address": "Địa chỉ Email",
+    "In order to go ahead, please sign in or create a new account!": "Để tiếp tục, vui lòng đăng nhập hoặc tạo một tài khoản mới!",
+    "Logout": "Đăng Xuất",
+    "Password": "Mật Khẩu",
+    "Sign In": "Đăng Nhập",
+    "SW360 is an open source software project that provides both a web application and a repository to collect, organize and make available information about software components It establishes a central hub for software components in an organization": "SW360 là một dự án phần mềm nguồn mở cung cấp cả ứng dụng web và kho lưu trữ để thu thập, sắp xếp và cung cấp thông tin có sẵn về các thành phần của phần mềm. Nó thiết lập một trung tâm trung tâm cho các thành phần phần mềm trong một tổ chức.",
+    "Start": "Bắt Đầu",
+    "Welcome to SW360!": "Chào mừng bạn đến với SW360!",
+    "Advance Search": "Tìm Kiếm Nâng Cao",
+    "Component Name": "Tên thành phần",
+    "Categories": "Thể loại",
+    "Component Type": "Loại thành phần",
+    "Group": "Nhóm",
+    "Languages": "Ngôn ngữ",
+    "Software Platforms": "Nền tảng phần mềm",
+    "Operating Systems": "Hệ điều hành",
+    "Vendors": "Nhà cung cấp",
+    "Main Licenses": "Giấy phép chính",
+    "Created By (Email)": "Được Tạo Bởi (Email)",
+    "Created On": "Được Tạo Vào",
+    "Search": "Tìm Kiếm",
+    "Add Component": "Thêm thành phần",
+    "Import SBOM": "Nhập SBOM",
+    "Export Spreadsheet": "Xuất bảng tính",
+    "Components Only": "Chỉ thành phần",
+    "Components With Releases": "Các thành phần có bản phát hành",
+    "Vendor": "Nhà cung cấp",
+    "Actions": "Hành động",
+    "Show": "Xem",
+    "entries": "mục",
+    "Internal": "Nội bộ",
+    "Inner Source": "Nguồn bên trong",
+    "Service": "Dịch vụ",
+    "Freeware": "Phần mềm miễn phí",
+    "Code Snippet": "Đoạn mã",
+    "The component cannot be deleted, since it contains releases Please delete the releases first": "Thành phần <strong>{name}</strong> không thể  bị xóa, vì nó chứa <strong>{releaseCount}</strong> bản phát hành. Vui lòng xóa các bản phát hành trước.",
+    "Error when processing!": "Lỗi khi xử lý!",
+    "Delete component success!": "Xóa thành phần thành công!",
+    "Delete Component": "Xóa Thành Phần",
+    "Do you really want to delete the component?": "Bạn có thực sự muốn xóa thành phần <strong>{name}</strong>?",
+    "Please comment your changes": "Hãy giải thích những thay đổi của bạn",
+    "Close": "Đóng",
+    "To": "Đến",
+    "Between": "Trong Khoảng"
+  }
+}
