+++ messages/zh.json
@@ -1,0 +1,50 @@
+{
+  "common": {
+    "Authentication failed Please try again": "认证失败，请重试。",
+    "Create Account": "创建账户",
+    "Email Address": "电子邮件地址",
+    "In order to go ahead, please sign in or create a new account!": "要继续，请登录或创建一个新帐户！",
+    "Logout": "登出",
+    "Password": "密码",
+    "Sign In": "登录",
+    "SW360 is an open source software project that provides both a web application and a repository to collect, organize and make available information about software components It establishes a central hub for software components in an organization": "SW360是一个开源软件项目，它提供了一个 Web 应用程序和一个存储库以收集，组织和提供有关软件组件的信息。它为组织中的软件组件建立了一个中心枢纽。",
+    "Start": "开始",
+    "Welcome to SW360!": "欢迎来到SW360！",
+    "Advance Search": "高级搜索",
+    "Component Name": "组件名称",
+    "Categories": "类别",
+    "Component Type": "组件类型",
+    "Group": "组",
+    "Languages": "语言",
+    "Software Platforms": "软件平台",
+    "Operating Systems": "操作系统",
+    "Vendors": "供应商",
+    "Main Licenses": "主许可证",
+    "Created By (Email)": "创建者（电子邮件）",
+    "Created On": "创建于",
+    "Search": "搜索",
+    "Add Component": "添加组件",
+    "Import SBOM": "导入SPDX BOM",
+    "Export Spreadsheet": "导出电子表格",
+    "Components Only": "仅组件",
+    "Components With Releases": "具有发行版本的组件",
+    "Vendor": "供应商",
+    "Actions": "操作",
+    "Show": "Show",
+    "entries": "entries",
+    "Internal": "内部",
+    "Inner Source": "内源",
+    "Service": "服务",
+    "Freeware": "免费软件",
+    "Code Snippet": "代码片段",
+    "The component cannot be deleted, since it contains releases Please delete the releases first": "组件 <strong>{name}</strong> 无法被删除，因为它包含发行版本 <strong>{releaseCount}</strong>。请先删除发行版本。",
+    "Error when processing!": "处理时出错！",
+    "Delete component success!": "删除组件成功！",
+    "Delete Component": "删除组件",
+    "Do you really want to delete the component?": "确定删除组件 <strong>{name}</strong>?",
+    "Please comment your changes": "请为变更添加注释",
+    "Close": "取消",
+    "To": "到",
+    "Between": "之间"
+  }
+}
