--- public/locales/en/common.json
@@ -1,48 +1,0 @@
-{
-  "Authentication failed. Please try again.": "Authentication failed. Please try again.",
-  "Create Account": "Create Account",
-  "Email Address": "Email Address",
-  "In order to go ahead, please sign in or create a new account!": "In order to go ahead, please sign in or create a new account!",
-  "Logout": "Sign Out",
-  "Password": "Password",
-  "Sign In": "Sign In",
-  "SW360 is an open source software project that provides both a web application and a repository to collect, organize and make available information about software components. It establishes a central hub for software components in an organization.": "SW360 is an open source software project that provides both a web application and a repository to collect, organize and make available information about software components. It establishes a central hub for software components in an organization.",
-  "Start": "Start",
-  "Welcome to SW360!": "Welcome to SW360!",
-  "Advance Search": "Advance Search",
-  "Component Name": "Component Name",
-  "Categories": "Categories",
-  "Component Type": "Component Type",
-  "Group": "Group",
-  "Languages": "Languages",
-  "Software Platforms": "Software Platforms",
-  "Operating Systems": "Operating Systems",
-  "Vendors": "Vendors",
-  "Main Licenses": "Main Licenses",
-  "Created By (Email)": "Created By (Email)",
-  "Created On": "Created On",
-  "Search": "Search",
-  "Add Component": "Add Component",
-  "Import SBOM": "Import SBOM",
-  "Export Spreadsheet": "Export Spreadsheet",
-  "Components Only": "Components Only",
-  "Components With Releases": "Components With Releases",
-  "Vendor": "Vendor",
-  "Actions": "Actions",
-  "Show": "Show",
-  "entries": "entries",
-  "Internal": "Internal",
-  "Inner Source": "Inner Source",
-  "Service": "Service",
-  "Freeware": "Freeware",
-  "Code Snippet": "Code Snippet",
-  "The component cannot be deleted, since it contains releases. Please delete the releases first.": "The component <strong>{{name}}</strong> cannot be deleted, since it contains <strong>{{releaseCount}}</strong> releases. Please delete the releases first.",
-  "Error when processing!": "Error when processing!",
-  "Delete component success!": "Delete component success!",
-  "Delete Component": "Delete Component",
-  "Do you really want to delete the component?": "Do you really want to delete the component <strong>{{name}}</strong>?",
-  "Please comment your changes": "Please comment your changes",
-  "Close": "Close",
-  "To": "To",
-  "Between": "Between"
-}
