--- public/locales/ja/common.json
@@ -1,48 +1,0 @@
-{
-  "Authentication failed. Please try again.": "認証に失敗しました。もう一度お試しください。",
-  "Create Account": "アカウントの作成",
-  "Email Address": "メールアドレス",
-  "In order to go ahead, please sign in or create a new account!": "先に進むためには、サインインするか、新しいアカウントを作成してください",
-  "Logout": "ログアウト",
-  "Password": "パスワード",
-  "Sign In": "サインイン",
-  "SW360 is an open source software project that provides both a web application and a repository to collect, organize and make available information about software components. It establishes a central hub for software components in an organization.": "SW360は、ソフトウェアコンポーネントに関する情報を収集、整理、利用可能にするためのウェブアプリケーションとリポジトリの両方を提供するオープンソースのソフトウェアプロジェクトです。SW360は、組織内のソフトウェアコンポーネントの一元化を行います。",
-  "Start": "開始",
-  "Welcome to SW360!": "SW360へようこそ！",
-  "Advance Search": "高度な検索",
-  "Component Name": "コンポーネント名",
-  "Categories": "カテゴリ",
-  "Component Type": "コンポーネントタイプ",
-  "Group": "グループ",
-  "Languages": "言語",
-  "Software Platforms": "ソフトウェアプラットフォーム",
-  "Operating Systems": "オペレーティングシステム",
-  "Vendors": "ベンダー",
-  "Main Licenses": "メインライセンス",
-  "Created By (Email)": "Created by (Email)",
-  "Created On": "作成日",
-  "Search": "検索",
-  "Add Component": "コンポーネントの追加",
-  "Import SBOM": "SBOMのインポート",
-  "Export Spreadsheet": "SBOMのインポート",
-  "Components Only": "コンポーネントのみ",
-  "Components With Releases": "リリースされたコンポーネント",
-  "Vendor": "ベンダー",
-  "Actions": "アクション",
-  "Show": "",
-  "entries": "件表示",
-  "Internal": "内部",
-  "Inner Source": "内部ソース",
-  "Service": "サービス",
-  "Freeware": "フリーウェア",
-  "Code Snippet": "コードスニペット",
-  "The component cannot be deleted, since it contains releases. Please delete the releases first.": "コンポーネント <strong>{{name}}</strong> は <strong>{{releaseCount}}</strong> を持つため消去できない．最初にリリースを消去してください",
-  "Error when processing!": "処理中にエラーが発生しました!",
-  "Delete component success!": "コンポーネントの削除が成功しました。",
-  "Delete Component": "コンポーネント削除",
-  "Do you really want to delete the component?": "本当にコンポーネント <strong>{{name}}</strong> を削除しますか？",
-  "Please comment your changes": "変更点にコメントしてください",
-  "Close": "キャンセル",
-  "To": "To",
-  "Between": "間"
-}
