+++ src/components/profile-dropdown/ProfileDropdown.tsx
@@ -1,0 +1,38 @@
+// Copyright (c) Helio Chissini de Castro, 2023. Part of the SW360 Frontend Project.
+
+// This program and the accompanying materials are made
+// available under the terms of the Eclipse Public License 2.0
+// which is available at https://www.eclipse.org/legal/epl-2.0/
+
+// SPDX-License-Identifier: EPL-2.0
+// License-Filename: LICENSE
+
+import NavDropdown from 'react-bootstrap/NavDropdown'
+import Image from 'next/image'
+import sw360ProfileIcon from '@/assets/images/profile.svg'
+import navbarStyles from '@/components/sw360-navbar/SW360Navbar.module.css'
+import { signOut } from 'next-auth/react';
+import { COMMON_NAMESPACE } from '@/object-types/Constants'
+import { useTranslations } from 'next-intl';
+
+const UserProfile = (
+    <Image
+        className={navbarStyles.profileImage}
+        src={sw360ProfileIcon}
+        alt='Profile'
+    />
+)
+
+function ProfileDropdown() {
+    const t = useTranslations(COMMON_NAMESPACE);
+    return (
+        <NavDropdown id="profileDropdown" title={UserProfile}>
+            <NavDropdown.Divider />
+            <NavDropdown.Item href='' onClick={() => signOut({callbackUrl: '/'})}>
+                {t('Logout')}
+            </NavDropdown.Item>
+        </NavDropdown>
+    )
+}
+
+export default ProfileDropdown
