+++ src/middleware.ts
@@ -8,17 +8,51 @@
 // SPDX-License-Identifier: EPL-2.0
 // License-Filename: LICENSE
 
-export { default } from "next-auth/middleware";
+import { NextRequest } from 'next/server';
+import { withAuth } from 'next-auth/middleware';
+import createIntlMiddleware from 'next-intl/middleware';
+import { LOCALES as locales } from './object-types/Constants';
+
+const publicPages = [
+    '/'
+];
+
+const intlMiddleware = createIntlMiddleware({
+    locales,
+    defaultLocale: 'en',
+    localePrefix: 'always'
+});
+
+const authMiddleware = withAuth(
+    // Note that this callback is only invoked if
+    // the `authorized` callback has returned `true`
+    // and not for pages listed in `pages`.
+    (req) => intlMiddleware(req),
+    {
+        callbacks: {
+            authorized: ({ token }) => token != null
+        },
+        pages: {
+            signIn: '/'
+        }
+    }
+);
+
+export default function middleware(req: NextRequest) {
+    const publicPathnameRegex = RegExp(
+        `^(/(${locales.join('|')}))?(${publicPages.join('|')})?/?$`,
+        'i'
+    );
+    const isPublicPage = publicPathnameRegex.test(req.nextUrl.pathname);
+
+    if (isPublicPage) {
+        return intlMiddleware(req);
+    } else {
+        return (authMiddleware as any)(req);
+    }
+}
 
 export const config = {
-    matcher: [
-        /*
-         * Match all request paths except for the ones starting with:
-         * - auth (auth routes)
-         * - _next/static (static files)
-         * - _next/image (image optimization files)
-         * - favicon.ico (favicon file)
-         */
-        '/((?!_next/static|_next/image|favicon.ico).*)',
-    ],
+    // Skip all paths that should not be internationalized
+    matcher: ['/((?!api|_next|.*\\..*).*)']
 };
