+++ src/object-types/Session.ts
@@ -1,6 +1,5 @@
 // Copyright (C) TOSHIBA CORPORATION, 2023. Part of the SW360 Frontend Project.
 // Copyright (C) Toshiba Software Development (Vietnam) Co., Ltd., 2023. Part of the SW360 Frontend Project.
-// Copyright (C) Siemens AG, 2023. Part of the SW360 Frontend Project.
 
 // This program and the accompanying materials are made
 // available under the terms of the Eclipse Public License 2.0
@@ -9,5 +8,15 @@
 // SPDX-License-Identifier: EPL-2.0
 // License-Filename: LICENSE
 
-export const SW360_API_URL: string | undefined  = process.env.NEXT_PUBLIC_SW360_API_URL;
-export const AUTH_TOKEN : string | undefined  = process.env.AUTH_TOKEN;
+export interface Session{
+  user? : {
+    access_token: string,
+    exp: number,
+    expires_in: number,
+    iat: number,
+    jti: string,
+    refresh_token: string,
+    scope: string
+    token_type: string,
+  }
+}
