+++ src/styles/auth.css
@@ -10,16 +10,7 @@
   License-Filename: LICENSE
 */
 
-.btn-primary {
-    background-color: #F7941E;
-    border-color: #F7941E;
-    color: #272833;
-}
 
-.btn-outline-primary {
-    border-color: #F7941E;
-    color: #F7941E;
-}
 
 .buttons a {
     margin-top: 1.75rem;
@@ -36,17 +27,7 @@
     padding-top: 0.59375rem;
 }
 
-.btn-primary:hover {
-    background-color: #f28809;
-    border-color: transparent;
-    color: #fff;
-}
 
-.btn-outline-primary:hover {
-    background-color: #fffdfb;
-    border-color: #F7941E;
-    color: #F7941E;
-}
 
 #portlet_sw360_portlet_welcome {
     margin: 2rem;
